exercise3-3.c
function expand(s1, s2) that ezpands shorthand notation like a-z in the string s1 into the equivalent complete list abc...xyz in s2. Allow for letter of either case and digits, and be prepared to handle cases like a-b-c and -a-z. Arrange that a leading or trailing - is taken literally.