Exercise 4-13: recursive reverse
Reverse works its way from both ends of the string towards the middle,
swapping characters at the two current indexes in each step, before
incrementing the start index and decrementing the end.

This means we have to keep track of these indexes and as a result the
reverse function had to be implemented using a helper function
recursereverse which takes these two indexes as parameters. The
function recursereverse calls itself recursively whilst reverse simply
starts the call to recursereverse with the first and last index
values.