Exercise 5-1: fix getint handle no digit after +/-
Getint would treat a + or - not followed by a digit as a valid
representation of zero.

Fix required addition of an extra variable to check if the value of
character after occurence of + or - was a digit. This is necessary
because if the next character is not a digit, we need to unget both this
character and the preceding + or -.