Exercise 7-1: lower / upper case conversion
Convert input to upper / lower case based on the name the program is
invoked with.

Solution: compare argv[0] (the name the program was invoked with) to
'lower' or 'upper' and store corresponding case conversion function in a
function pointer. Then call this function on each character read from
input and print to standard output.